Actual -> Background image processing complete
